connect register photogrpher to mobile profile, and add prpgress bar animation
