+++ lib/feature/profile/register_photographer/view/register_photographer_view.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:foto_in/feature/profile/register_photographer/widget/mobile/register_photograph_mobile.dart';
 import 'package:foto_in/feature/profile/register_photographer/widget/register_photographer_wiget.dart';
+import 'package:responsive_builder/responsive_builder.dart';
 
 class RegisterPhotographer extends StatelessWidget {
   static const String routeName = '/register_photographer';
@@ -8,6 +10,13 @@ class RegisterPhotographer extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const RegisterPhotographerWidget();
+    // return const RegisterPhotographerWidget();
+    return ResponsiveBuilder(builder: (context, sizingInformation) {
+      if (sizingInformation.deviceScreenType == DeviceScreenType.desktop) {
+        return const RegisterPhotographerWidget();
+      } else {
+        return const ProgressPage();
+      }
+    });
   }
 }
