+++ lib/feature/profile/widget/mobile/profile_mobile.dart
@@ -7,7 +7,9 @@ import 'package:flutter/widgets.dart';
 import 'package:foto_in/core/styles/colors.dart';
 import 'package:foto_in/core/styles/typography.dart';
 import 'package:foto_in/feature/home/presentation/widgets/mobile/register_photographer_banner.dart';
+import 'package:foto_in/feature/profile/register_photographer/view/register_photographer_view.dart';
 import 'package:foto_in/feature/profile/register_photographer/widget/profile_form_title.dart';
+import 'package:foto_in/feature/profile/widget/mobile/profile_menu_item.dart';
 
 class ProfileMobile extends StatelessWidget {
   const ProfileMobile({super.key});
@@ -116,7 +118,14 @@ class ProfileMobile extends StatelessWidget {
                   ),
                   Padding(
                     padding: const EdgeInsets.symmetric(horizontal: 16),
-                    child: RegisterPhotographerBannerMobile(),
+                    child: RegisterPhotographerBannerMobile(
+                      onPressed: () {
+                        Navigator.pushNamed(
+                          context,
+                          RegisterPhotographer.routeName,
+                        );
+                      },
+                    ),
                   ),
                   Container(
                     margin: const EdgeInsets.symmetric(
@@ -147,61 +156,6 @@ class ProfileMobile extends StatelessWidget {
                   ),
                 ],
               ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class ProfileMenuItem extends StatelessWidget {
-  const ProfileMenuItem({
-    super.key,
-    required this.leadingIcon,
-    required this.title,
-    required this.trailingIcon,
-    this.onTap,
-  });
-
-  final IconData leadingIcon;
-  final String title;
-  final IconData trailingIcon;
-  final Function()? onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: onTap,
-      child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 20),
-        decoration: const BoxDecoration(
-          border: Border(
-            bottom: BorderSide(
-              color: Color(0xffE9E9E9),
-              width: 1,
-            ),
-          ),
-        ),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Row(
-              children: [
-                Icon(
-                  leadingIcon,
-                ),
-                SizedBox(
-                  width: 16,
-                ),
-                Text(
-                  "Ubah Password",
-                  style: FotoInHeadingTypography.xxSmall(),
-                ),
-              ],
-            ),
-            Icon(
-              trailingIcon,
             ),
           ],
         ),
